+++ frontend/src/components/modal/FolderModal/FolderModal.stories.tsx
@@ -1,0 +1,10 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { FolderModal } from './FolderModal';
+
+type T = typeof FolderModal
+
+export default {
+  component: FolderModal,
+} satisfies Meta<T>;
+
+export const Default: StoryObj<T> = {};
